// src/app/core/util/enums.ts
// enum.ts

export enum SeverityEnum {
  SEVERITY_1 = 'SEVERITY_1',
  SEVERITY_2 = 'SEVERITY_2',
  SEVERITY_3 = 'SEVERITY_3',
  SEVERITY_4 = 'SEVERITY_4',
}

export enum ProductEnum {
  KORE_AI_XO_PLATFORM = 'KORE_AI_XO_PLATFORM',
  SMART_ASSIST = 'SMART_ASSIST',
  BANK_ASSIST = 'BANK_ASSIST',
  HEALTH_ASSIST = 'HEALTH_ASSIST',
  HR_ASSIST = 'HR_ASSIST',
  IT_ASSIST = 'IT_ASSIST',
  SEARCH_ASSIST = 'SEARCH_ASSIST',
  AGENT_ASSIST_SMART_ASSIST = 'AGENT_ASSIST_SMART_ASSIST',
  WORK_ASSIST_NOT_SUPPORTED_YET = 'WORK_ASSIST_NOT_SUPPORTED_YET',
  KORA = 'KORA',
  NONE = 'NONE',
  AGENT_DESKTOP_DEPRECATED_USE_SMART_ASSIST = 'AGENT_DESKTOP_DEPRECATED_USE_SMART_ASSIST',
  GALE = 'GALE',
}

export enum EnvironmentEnum {
  NO_AFFECTED_ENVIRONMENT = 'NO_AFFECTED_ENVIRONMENT',
  NON_PROD_US_SAAS = 'NON_PROD_US_SAAS',
  NON_PROD_AU_SAAS = 'NON_PROD_AU_SAAS',
  NON_PROD_DE_SAAS = 'NON_PROD_DE_SAAS',
  NON_PROD_EU_SAAS = 'NON_PROD_EU_SAAS',
  NON_PROD_JP_SAAS = 'NON_PROD_JP_SAAS',
  NON_PROD_IND_SAAS = 'NON_PROD_IND_SAAS',
  NON_PROD_ON_PREMISE_PRIVATE_SAAS = 'NON_PROD_ON_PREMISE_PRIVATE_SAAS',
  PROD_US_SAAS = 'PROD_US_SAAS',
  PROD_AU_SAAS = 'PROD_AU_SAAS',
  PROD_DE_SAAS = 'PROD_DE_SAAS',
  PROD_EU_SAAS = 'PROD_EU_SAAS',
  PROD_JP_SAAS = 'PROD_JP_SAAS',
  PROD_IND_SAAS = 'PROD_IND_SAAS',
  PROD_ON_PREMISE_PRIVATE_SAAS = 'PROD_ON_PREMISE_PRIVATE_SAAS',
}

export enum InstallationTypeEnum {
  ON_PREMISE = 'ON_PREMISE',
  SAAS_MULTI_TENANT = 'SAAS_MULTI_TENANT',
  SAAS_PRIVATE = 'SAAS_PRIVATE',
}

export enum MessageTypeEnum {
  CLIENT = 'CLIENT',
  VENDOR = 'VENDOR',
}

export enum StatusEnum {
  AWAITING_REPLY = 'AWAITING_REPLY',
  SOLVED = 'SOLVED',
  CLOSED = 'CLOSED',
  OPEN = 'OPEN',
}

export enum SupportRequestTypeEnum {
  INCIDENT = 'INCIDENT',
  QUESTION = 'QUESTION',
  SERVICE_REQUEST = 'SERVICE_REQUEST',
  ROOT_CAUSE_ANALYSIS = 'ROOT_CAUSE_ANALYSIS',
  ENHANCEMENT = 'ENHANCEMENT',
  PROVISIONING = 'PROVISIONING',
  REPORTED_VULNERABILITY = 'REPORTED_VULNERABILITY',
}